Refactor axis view to own module
