+++ lib/chart/line/axis_svg.ex
@@ -1,13 +1,14 @@
 defmodule Chart.Line.AxisSvg do
   @moduledoc false
 
-  alias Chart.Line.View
+  alias Chart.Line.{AxisView, View}
   use Phoenix.HTML
 
   def render(settings) do
     assigns =
       settings
       |> Map.put(:view, View)
+      |> Map.put(:axis_view, AxisView)
 
     ~E"""
     <g class="axis">
@@ -15,11 +16,11 @@ defmodule Chart.Line.AxisSvg do
       <% settings_ax = settings[axis] %>
       <% minor_ticks = settings_ax.minor_ticks %>
       <defs>
-        <% {x1, y1} = @view.set_axis_tick(minor_ticks.length, vector) %>
-        <% css_id = @view.css_id_axis_minor_line_ticks(axis) %>
+        <% {x1, y1} = @axis_view.set_axis_tick(minor_ticks.length, vector) %>
+        <% css_id = @axis_view.css_id_axis_minor_line_ticks(axis) %>
         <line id="<%= css_id %>" x1="<%= x1 %>" y1="<%= y1 %>" />
       </defs>
-      <g class="<%= @view.css_class_axis_minor_ticks(axis) %>">
+      <g class="<%= @axis_view.css_class_axis_minor_ticks(axis) %>">
         <%= for position <- minor_ticks.positions do %>
           <use xlink:href="#<%= css_id %>"
             transform="<%= @view.translate_axis_ticks(settings_ax.line, minor_ticks.length, minor_ticks.gap, position, vector) %>" />
@@ -28,11 +29,11 @@ defmodule Chart.Line.AxisSvg do
 
       <% major_ticks = settings_ax.major_ticks %>
       <defs>
-        <% {x1, y1} = @view.set_axis_tick(major_ticks.length, vector) %>
-        <% css_id = @view.css_id_axis_major_line_ticks(axis) %>
+        <% {x1, y1} = @axis_view.set_axis_tick(major_ticks.length, vector) %>
+        <% css_id = @axis_view.css_id_axis_major_line_ticks(axis) %>
         <line id="<%= css_id %>" x1="<%= x1 %>" y1="<%= y1 %>" />
       </defs>
-      <g class="<%= @view.css_class_axis_major_ticks(axis) %>">
+      <g class="<%= @axis_view.css_class_axis_major_ticks(axis) %>">
         <%= for position <- major_ticks.positions do %>
           <use xlink:href="#<%= css_id %>"
             transform="<%= @view.translate_axis_ticks(settings_ax.line, major_ticks.length, major_ticks.gap, position, vector) %>" />
@@ -40,9 +41,9 @@ defmodule Chart.Line.AxisSvg do
       </g>
 
       <% major_ticks_text = @view.axis_ticks_label(settings_ax.line, settings_ax.major_ticks_text, vector) %>
-      <g class="<%= @view.css_class_axis_major_ticks_label(axis) %>">
+      <g class="<%= @axis_view.css_class_axis_major_ticks_label(axis) %>">
         <%= for {pos_x, pos_y, label} <- major_ticks_text do %>
-          <text class="<%= @view.css_class_axis_tick_label(axis) %>" x="<%= pos_x %>" y="<%= pos_y %>"
+          <text class="<%= @axis_view.css_class_axis_tick_label(axis) %>" x="<%= pos_x %>" y="<%= pos_y %>"
             alignment-baseline="middle" text-anchor="middle"
           ><%= label %></text>
         <% end %>
@@ -54,7 +55,7 @@ defmodule Chart.Line.AxisSvg do
 
       <g class="axis-label">
         <% {pos_x, pos_y} = settings_ax.label.position %>
-          <text id="<%= @view.css_id_axis_label(axis) %>" x="<%= pos_x %>" y="<%= pos_y %>"
+          <text id="<%= @axis_view.css_id_axis_label(axis) %>" x="<%= pos_x %>" y="<%= pos_y %>"
             alignment-baseline="middle" text-anchor="middle"
           ><%= settings_ax.label.text %></text>
       </g>
